+++ app/src/main/java/com/techcourse/HandlerMappingRegistry.java
@@ -1,14 +1,15 @@
 package com.techcourse;
 
+import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.Objects;
 
 public class HandlerMappingRegistry {
 
-    private final List<HandlerMapping> handlerMappings;
+    private final List<com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping> handlerMappings;
 
-    public HandlerMappingRegistry(List<HandlerMapping> handlerMappings) {
+    public HandlerMappingRegistry(List<com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping> handlerMappings) {
         this.handlerMappings = handlerMappings;
         handlerMappings.forEach(HandlerMapping::initialize);
     }
