+++ app/src/main/java/com/techcourse/handlermapping/ManualHandlerMapping.java
@@ -2,7 +2,7 @@ package com.techcourse.handlermapping;
 
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.asis.ForwardController;
-import com.techcourse.HandlerMapping;
+import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
 import com.techcourse.controller.LoginController;
 import com.techcourse.controller.LoginViewController;
 import com.techcourse.controller.LogoutController;
